project_mercury/rtm/src/main/java/com/smartsparrow/rtm/subscription/courseware/pathwayreordered/PathwayReOrderedRTMConsumable.java
package com.smartsparrow.rtm.subscription.courseware.pathwayreordered;

import com.smartsparrow.rtm.message.RTMClientContext;
import com.smartsparrow.rtm.subscription.courseware.ActivityRTMSubscription;
import com.smartsparrow.rtm.subscription.courseware.message.PathwayReOrderedBroadcastMessage;
import com.smartsparrow.rtm.subscription.data.AbstractRTMConsumable;
import com.smartsparrow.pubsub.data.RTMEvent;

/**
 * This RTM consumable describes a pathway reordered event
 */
public class PathwayReOrderedRTMConsumable extends AbstractRTMConsumable<PathwayReOrderedBroadcastMessage> {

    private static final long serialVersionUID = -101147421067995506L;

    public PathwayReOrderedRTMConsumable(RTMClientContext rtmClientContext, PathwayReOrderedBroadcastMessage content) {
        super(rtmClientContext, content);
    }

    @Override
    public RTMClientContext getRTMClientContext() {
        return rtmClientContext;
    }

    @Override
    public String getName() {
        return String.format("author.activity/%s/%s", content.getActivityId(), getRTMEvent().getName());
    }

    @Override
    public String getSubscriptionName() {
        return ActivityRTMSubscription.NAME(content.getActivityId());
    }

    @Override
    public RTMEvent getRTMEvent() {
        return new PathwayReOrderedRTMEvent();
    }

    @Override
    public boolean equals(Object o) {
        return super.equals(o);
    }

    @Override
    public int hashCode() {
        return super.hashCode();
    }
}
